requirements.txt
flask
flask_restful
pillow
flask_dropzone
flask_sqlalchemy
flask_bcrypt
flask_login
flask_uploads
werkzeug==0.16.0
flask_wtf